requirements.txt
google-genai
google-adk
fastapi
uvicorn
python-dotenv
websockets
